restructured the packages and created a setup script
