+++ setup.py
@@ -1,0 +1,15 @@
+from distutils.core import setup
+from setuptools import find_packages
+
+requires = ["six>=1.10.0"]
+
+if __name__ == "__main__":
+    setup(
+        name="vusic-models",
+        packages=find_packages(),
+        author="DAWG",
+        author_email="DAWG@vusic.com",
+        install_requires=requires,
+        description="The API to fetch all packages",
+        include_package_data=True,
+    )
