--- test_sample.py
@@ -1,7 +1,0 @@
-# content of test_sample.py
-def inc(x):
-    return x + 1
-
-
-def test_answer():
-    assert inc(4) == 5
